Move git graph from models to modules/graph
