+++ routers/repo/commit.go
@@ -14,6 +14,7 @@ import (
 	"code.gitea.io/gitea/modules/charset"
 	"code.gitea.io/gitea/modules/context"
 	"code.gitea.io/gitea/modules/git"
+	"code.gitea.io/gitea/modules/gitgraph"
 	"code.gitea.io/gitea/modules/log"
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/services/gitdiff"
@@ -99,7 +100,7 @@ func Graph(ctx *context.Context) {
 
 	page := ctx.QueryInt("page")
 
-	graph, err := models.GetCommitGraph(ctx.Repo.GitRepo, page)
+	graph, err := gitgraph.GetCommitGraph(ctx.Repo.GitRepo, page)
 	if err != nil {
 		ctx.ServerError("GetCommitGraph", err)
 		return
